stream: check for decoder.end before calling it
In node <= v0.8.x, the stream_decoder.end() function was not yet implemented,
so attempting to invoke the undefined function leads to an error being thrown.

We want to attempt to keep these lib/_stream_* files backwards compatible with
at least v0.8.x for a little while, since these files get back-ported to the
isaacs/readable-stream module which is intended for older versions of node to use.