repl: treat "Assignment to constant variable" as a syntax error
Fixes #6132.

Adding a new `repl-harmony` test file here because adding the
`--use_strict --harmony` flags on the main repl test file was causing
lots of unrelated failures, due to global variable assignments and things
like that. This new test file is based off of the original repl.js test
file, but has a lot of the tests stripped out. A test case for this commit
is included though.